exam-generator/requirements.txt
google-generativeai>=0.8.0
pydantic>=2.9.0
requests>=2.32.0
python-dotenv>=1.0.0
